DEV-1726 Make test use ENV var of OC URL
A test failed because it expects a fixed URL of a certain OpenShift
environment. However, the actual code makes use of the ENV var OC_URL. When
running a Jenkins build, that OC_URL gets set with the URL of the OpenShift
environment it is running in. This means that if the test runs on a different
environment (OC cluster) than specified in the test, the test fails.

The test now uses/expects the OC_URL ENV var, so that it is independent of
the actual OC cluster.